NPM 공통함수 만들고 사용하기
