+++ app.js
@@ -1,27 +1,16 @@
-
-/*    
 const express = require('express')
+const myCommon = require('ryong-common-fun')
 const app = express()
 const port = 3000
 
 app.get('/', (req, res) => {
-	res.send('Hello World')
+  const a = 1
+  const b = 10
+  let myJson = { "Hello": "World", "ADD": myCommon.add(a, b) }
+
+  res.json(myJson)
 })
 
 app.listen(port, () => {
-	console.log(`Example app listening on port ${port}`)
+  console.log(`Example app listening on port ${port}`)
 })
-*/
-//Route parameters
-const express = require('express')
-const router = express.Router();
-
-router.get('/artists/:id', function (req, res) {
-	console.log("id는 " + req.params.id + "입니다")
-	res.send("id : " + req.params.id)
-});
-
-//여러개도 가능
-router.get('/artists/:id/company/:company', function(req,res){
-	res.send("id : " + req.params.id + " 회사 : " + req.params.company)
-});
