+++ package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/INAUGURATE-Ryong/Ryong-express#readme",
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "ryong-common-fun": "^0.2.1"
   }
 }
